Refactor dashboard fixtures
-  Split the big file into multiple smaller ones
-  Dashboard.fixtures now contains just some top-level, simple stuff
-  SimpleDashboard.fixtures.ts - everything related to the 'simple' reference dashboard
-  ComplexDashboard.fixtures.ts - everything related to the 'complex' reference dashboard
-  FilteredDashboards.fixtures.ts - everything related to reference dashboards used to verify filter queries
-  Layout.fixtures.ts - various small mocks & factories for objects that play in the dashboard layout
-  CatalogAvailability.fixtures.ts - functions to use to mock catalog availability

JIRA: RAIL-3492
