+++ libs/sdk-ui-dashboard/src/model/commandHandlers/filterContext/attributeFilter/tests/addAttributeFilterHandler.test.ts
@@ -3,7 +3,7 @@ import { ReferenceLdm } from "@gooddata/reference-workspace";
 import { addAttributeFilter } from "../../../../commands";
 import { DashboardTester, preloadedTesterFactory } from "../../../../tests/DashboardTester";
 import { selectFilterContextAttributeFilters } from "../../../../state/filterContext/filterContextSelectors";
-import { SimpleDashboardIdentifier } from "../../../../tests/Dashboard.fixtures";
+import { SimpleDashboardIdentifier } from "../../../../tests/fixtures/SimpleDashboard.fixtures";
 
 describe("addAttributeFilterHandler", () => {
     let Tester: DashboardTester;
