+++ libs/sdk-ui-dashboard/src/model/commandHandlers/filterContext/attributeFilter/tests/changeAttributeFilterSelectionHandler.test.ts
@@ -2,7 +2,7 @@
 import { changeAttributeFilterSelection } from "../../../../commands";
 import { DashboardTester, preloadedTesterFactory } from "../../../../tests/DashboardTester";
 import { selectFilterContextAttributeFilters } from "../../../../state/filterContext/filterContextSelectors";
-import { SimpleDashboardIdentifier } from "../../../../tests/Dashboard.fixtures";
+import { SimpleDashboardIdentifier } from "../../../../tests/fixtures/SimpleDashboard.fixtures";
 
 describe("changeAttributeFilterSelectionHandler.test", () => {
     let Tester: DashboardTester;
