+++ libs/sdk-ui-dashboard/src/model/commandHandlers/filterContext/attributeFilter/tests/moveAttributeFilterHandler.test.ts
@@ -2,7 +2,7 @@
 import { moveAttributeFilter } from "../../../../commands";
 import { DashboardTester, preloadedTesterFactory } from "../../../../tests/DashboardTester";
 import { selectFilterContextAttributeFilters } from "../../../../state/filterContext/filterContextSelectors";
-import { SimpleDashboardIdentifier } from "../../../../tests/Dashboard.fixtures";
+import { SimpleDashboardIdentifier } from "../../../../tests/fixtures/SimpleDashboard.fixtures";
 
 describe("moveAttributeFilterHandler", () => {
     let Tester: DashboardTester;
