+++ libs/sdk-ui-dashboard/src/model/commandHandlers/filterContext/dateFilter/tests/changeDateFilterSelectionHandler.test.ts
@@ -2,7 +2,7 @@
 import { changeDateFilterSelection, clearDateFilterSelection } from "../../../../commands";
 import { DashboardTester, preloadedTesterFactory } from "../../../../tests/DashboardTester";
 import { selectFilterContextDateFilter } from "../../../../state/filterContext/filterContextSelectors";
-import { SimpleDashboardIdentifier } from "../../../../tests/Dashboard.fixtures";
+import { SimpleDashboardIdentifier } from "../../../../tests/fixtures/SimpleDashboard.fixtures";
 
 describe("changeDateFilterSelectionHandler", () => {
     let Tester: DashboardTester;
