+++ libs/sdk-ui-dashboard/src/model/commandHandlers/layout/tests/moveSectionItemHandler.test.ts
@@ -1,12 +1,7 @@
 // (C) 2021 GoodData Corporation
 
 import { DashboardTester, preloadedTesterFactory } from "../../../tests/DashboardTester";
-import {
-    ComplexDashboardIdentifier,
-    ComplexDashboardWithReferences,
-    SimpleDashboardIdentifier,
-    TestCorrelation,
-} from "../../../tests/Dashboard.fixtures";
+import { TestCorrelation } from "../../../tests/fixtures/Dashboard.fixtures";
 import { moveSectionItem, undoLayoutChanges } from "../../../commands";
 import {
     DashboardCommandFailed,
@@ -14,6 +9,11 @@ import {
     DashboardLayoutSectionItemMoved,
 } from "../../../events";
 import { selectLayout } from "../../../state/layout/layoutSelectors";
+import { SimpleDashboardIdentifier } from "../../../tests/fixtures/SimpleDashboard.fixtures";
+import {
+    ComplexDashboardIdentifier,
+    ComplexDashboardWithReferences,
+} from "../../../tests/fixtures/ComplexDashboard.fixtures";
 
 describe("move layout section item handler", () => {
     describe("for any dashboard", () => {
