+++ libs/sdk-ui-dashboard/src/model/queryServices/tests/queryInsightDateDatasets.test.ts
@@ -1,19 +1,20 @@
 // (C) 2021 GoodData Corporation
 import { DashboardTester, preloadedTesterFactory } from "../../tests/DashboardTester";
-import {
-    createTestInsightItem,
-    EmptyDashboardIdentifier,
-    MockAvailabilityWithDifferentRelevance,
-    MockAvailabilityWithSameRelevance,
-    PivotTableWithDateFilter,
-    PivotTableWithRowAndColumnAttributes,
-    TestSectionHeader,
-    TreemapWithOneMeasureAndViewByDateAndSegmentByDate,
-} from "../../tests/Dashboard.fixtures";
+import { EmptyDashboardIdentifier } from "../../tests/fixtures/Dashboard.fixtures";
 import { addLayoutSection, loadDashboard } from "../../commands";
 import { InsightDateDatasets, queryDateDatasetsForInsight } from "../../queries";
 import { IInsight, insightRef } from "@gooddata/sdk-model";
 import { ICatalogDateDataset } from "@gooddata/sdk-backend-spi";
+import {
+    MockAvailabilityWithDifferentRelevance,
+    MockAvailabilityWithSameRelevance,
+} from "../../tests/fixtures/CatalogAvailability.fixtures";
+import {
+    PivotTableWithDateFilter,
+    PivotTableWithRowAndColumnAttributes,
+    TreemapWithOneMeasureAndViewByDateAndSegmentByDate,
+} from "../../tests/fixtures/Insights.fixtures";
+import { createTestInsightItem, TestSectionHeader } from "../../tests/fixtures/Layout.fixtures";
 
 function datasetsDigest(
     datasets: ReadonlyArray<ICatalogDateDataset | undefined>,
